fix: aldryn config parsing
The result of clean gets used to display forms as well, so it would end
up with a string-ification of the values after clean
("[u'item1', u'item2']") instead of the string with commas.

The new implementation cleans the comma separated items to always have
a space after the comma. Conversion to an actual python list happens
while generating the settings.